requirements.txt
streamlit
youtube-transcript-api
langchain
langchain-community
langchain-huggingface
huggingface_hub
tiktoken
faiss-cpu
python-dotenv
